More scaffolding and initial work on LiveState.
@thusc I've been working on the right abstractions to have a state
displayed via a terminal UI; more on that soon.

So far, I've made the "output" parameterised. This means we can switch
outputs from Brick to something else. However, how to wire this
elegantly is another question (I'm working on that).

We can, in the future, also add a new output "Repl" or something else,
which, as we discussed, will modify the state via a Repl.

For now, I've gone ahead with Brick; which seems like a bit of initial
work, but it seems to fit the bill.