Tie it all together.
@thusc we needed a minor adjustment in the
lib/src/Prototype/Backend/InteractiveState/Repl.hs file; the function
to run the `m` into an `IO` needed to return some `Either err a` in IO
since we'd like to safely map all errors.

The REPL starts up, but it hangs at the first error. The parsing also
seems to be a bit off at this time.

However, this, /does/ implement things as we'd like: the repl seems a
very realistic instant state modification medium.

I'd be interested in getting feedback/questions here.