Call the module "Main" instead of ExampleMain.
@thusc this is the cause: https://stackoverflow.com/a/27591340/1360368

And this is a good example why Cabal is not great to work with.
Pretty much anywhere else in a project, the module name must match the
dir. path and the file name, except for the main module, where the
file can be called anything, but the module needs to be called Main.

This is pretty nasty; and after all these years, I still get bitten by
this now and then.