More attempts to fix forms.
I still seem to be having quite a few issues with even the simplest of
forms. @thusc, I've followed your advice, but I cannot get the signup
page to work. On verifying the html, the form looks correct to me.
However; the signup/create EP complains that it cannot find the form
parameters it expects. For example; `PasswordConfirmation` is a valid
field in the form and it is created. But that field is not visible to
servant as post-data.

I'll keep trying to fix this but I've lost quite a bit of time on
this, again. I guess I really need to brush up my HTML.