Another rather frustrating attempt to get the form to work.
@thusc I've now tried the product type approach, expecting there to be
some idiosyncracies in the way forms work.

This works, but its pretty frustrating to expect creating a full
product type on each form submission. I'd have liked to process fields
individually, much like what we can do with query parameters.

Using query params here is not an option since passwords will get
exposed on the URL. That is not an issue over SSL but it will sooner
or later end up in logs in common practice.

We also may have some issues with our logging; I'll check that next weekend.