Make the server thread exit when the REPL exits.
Instead of using `concurrently`, we simply use `forkIO` to run the
server thread. As soon as the main thread exits, so does the server
thread. (And the main thread is used by the REPL.)

Additionaly, the code is re-structured a bit to expose more of the
similarities between the replProcess and serverProcess (the start/end
functions).

An unused `settings` variable has been deleted.

We also show the port the server is listening on.