Replace Readline by Haskeline.
The main reason to do so was to solve:

https://github.com/asheshambasta/prototype-hs/issues/6

where it seemed that the output of the REPL was mixed up with the output
of the server. Something similar was also happening without the server
code but when using the REPL within GHCi.

I no longer see the same behavior after the change.

It's also interesting to note that Haskeline is BSD-3 while Readline is
GPL (since it's a binding to GNU Readline) and has many more reverse
dependencies on Hackage:

https://packdeps.haskellers.com/reverse/haskeline

Note that the previous loop code was using an argument to decide whether
to continue or not (ReplContinue) but this didn't seem used.