[WIP] Work on user profiles.
While type-safe, servant authentication is a pretty good example of
over-engineering. I, to this day, have a hard time keeping track of
contexts etc. with authentication. Wiring up the server has also never
been a pleasant experience for me.

This is basically close to done, but as we wire in the Private server,
we get a mysterious error:

```
src/Prototype/Example/Server.hs:59:3-43: error:
    • Could not deduce (Srv.ToJWT
                          Prototype.Example.Data.User.UserProfile)
        arising from a use of ‘serveWithContext’
      from the context: Pub.PublicServerC m
        bound by the type signature for:
                   exampleApplication :: forall (m :: * -> *).
                                         Pub.PublicServerC m =>
                                         (forall x. m x -> Handler x) -> Application
        at src/Prototype/Example/Server.hs:(53,1)-(57,20)
    • In the expression: serveWithContext exampleProxy ctx
      In the expression:
        serveWithContext exampleProxy ctx
          $ hoistServerWithContext
              exampleProxy settingsProxy handlerNatTrans exampleT
      In an equation for ‘exampleApplication’:
          exampleApplication handlerNatTrans
            = serveWithContext exampleProxy ctx
                $ hoistServerWithContext
                    exampleProxy settingsProxy handlerNatTrans exampleT
            where
                exampleProxy = Proxy @Example
                settingsProxy = Proxy @ServerSettings
                ctx = undefined
   |
59 |   Servant.serveWithContext exampleProxy ctx $ hoistServerWithContext exampleProxy
   |   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```

For some reason, we end up needing an instance of ToJWT of the
UserProfile.
We don't want this since we don't want to store the user in the
user-cookie, just the ID is sufficient.

I'm still battling with this but this has costed some precious time.

@thusc.