More human friendly login: don't require user-ids.
@thusc as we discussed, this makes the login process also work with
the UserName, if it yields an unambiguous set of users.

As a less secure option, on reaching this unambiguity, we can still
attempt to authenticate amongst the found users and return the first
that succeeds, but that is far less secure.

However, for a simplistic application like this, it should be fine.
Thoughts?