Use withAsync.
@noteed, as advised, I've now switched the boot process to clean up
orphan threads once the main server process has exited.
We can, possibly, allow different behaviours based on a configuration
parameter in which we can "unlink" the server and repl processes and
wait for the repl to exit when the server process has exited.

However, lets do that when it seems useful, IMO.

This completes the PR advised changes. I'll leave it to you to merge
the PR when you do a final quick review.