Increase the size of the disk for Digital Ocean.
@asheshambasta This should resolve a problem you experienced when build
the Digital Ocean image. (I thought previously it was when building the
QEMU image.)

I thought the default value was 'auto' so I'm a bit surprised I had to
manually set that option. That being said, I think there is something
weird goind on: for instance I think this is including GHC in the image,
which shouldn't be the case.