Now, wire everything in for DB serialisation.
DB state can now be read from a file and written to a file on
application exit. The current support is rudimentary;

1. IFF a file is provided:
If the file exists & is non-empty, try to read the DB state from it.
If decoding fails, report these errors and exit.

If the file is absent, or is empty, start with an empty state, but
write the eventual state to the file on powerdown.

2. IFF no file is provided:
Don't read any initial state from a file, nor write any initial state
to file.