Make booting from a non-existing file an error.
Same if the file is empty.

The previous behavior was to silently continue, then save the state to
the given state file at powerdown. I guess that seemed like a good idea
to make the life of the users simpler, but I think it actually makes it
a bit more error-prone (e.g. a mistyped file would end-up created, and
the server would be running with a different state than the user
intended).

@asheshambasta What do you think of the fact that the following two
errors look quite different ?

    *Main> :main --state non-existing.json user get alice
    *** Exception: FileDoesntExistErr "non-existing.json"

    *Main> :main --state empty.json user get alice
    *** Exception: httpStatus = Status {statusCode = 500, statusMessage = "Internal Server Error"} userMessage = Error in $: not enough input