Add notes about deploying smartcoop.sh at DO.
This commit introduces four scripts and a new "toplevel" attribute in
default.nix.

The four scripts come from https://github.com/noteed/nix-notes and are
used to upload and import an image into DO, create a Droplet, and deploy
changes to it.

Changes are built as a new "toplevel" attributes, which represent all
the Nix store paths that must be uploaded to the machine and then
activated.