Move the documentation to a subdir.
The subdir is content/.

The goal is to organise the files so that `documentation/index.html`
becomes `documentation.html`.

With such a scheme, the static serving feature of our application can be
changed to map a URL such as `documentation` or `documentation/sitemap` to
points to `documentation.html`, or `documentation/sitemap.html`.