Use smartcoop design for the welcome page.
@asheshambasta I find myself removing the use of Page Public Void or
Page Authd UserProfile, or the automatic navigation bar provided by the
IsNavbarContent class:

Could you have a look if Public/Authd could be re-added to the route
type signature ? I'm also wondering if this proves really something, and
we already have UserAuthentication guarding the route.

For the automatic navigation bar with IsNavbarContent, I think it's best
to see how the application evolves before trying to over-structure
things too much. For instance, I think that some screens would not show
such a navigation bar, even when logged in. For instance I'm thinking to
some kind of "fullscreen" page with a 'X' close button.