Wire in TOS consent in to the user.
The forms etc. now accept the values in the checkbox to record if the
user agrees with the terms of service.

Currently, this isn't enforced, but can be done.

I'd postpone that to the error modes fix since this will be a good
test use case.

For now this can be merged as a smaller PR that starts recording the
checked/unchecked values.