Add an `as` command to `cty run`.
Normally, `cty` supports a `--user` option, which is set by default to
the UNIX login.

Within a script executed by `cty run`, such an option is not recognized.
(It is recognized on the command-line though, e.g. `cty --user alice
run`.)

This change allows to set that user for the rest of the script with a
line such as

    as alice