Quick proof of concept: use a Counter type for generating IDs.
@noteed we use a generalised class for all the ID types. In this proof
of concept, I've only replaced the business-ids.

--

The core idea is to have a newtype that wraps around a counter value.
Then we have some notion of `m` that wraps around counter values. And
we provide one generalised way to be able to generate new counter
values, bump values, and instantiate them to a desired value.

This can be a drop-in replacement of the repetition you mentioned; and
can provide significant code size reduction, especially as we add more
entities operating with IDs.

This is just a proof of concept, so I'd be curious to know what you
think.