Add a random generator state to Curiosity state.
By keeping the state of the pseudo random number generator within the
complete system state, we can deterministically generate those numbers,
which is important to reproduce faithfully the test scenarios.

In a production system, we could let the initial seed be different (and
probably stop saving that internal state to the real production
database).