Full implementation:
We basically create 2 Applications, and we run the serveNamespace
function as a "Test". That is, we will ignore its results if no
namespace is found, or if it is found; but we'll use Either to indicate
a hit or a miss.

If we do have a miss; we fallback to the Documentation Application.

Several such fallbacks can be chained together.

--

@noteed: what we're trying to do here is possibly _very_ hard if not
impossible in servant.

I also think route-failure-fallbacks (as in the recent features) are
something that I possibly wouldn't advise in production. There may be
cases where retries/fallbacks are eating through resources without a
clear way of knowing why.

However, that is possibly a subjective matter.