Add some safe instantiation and closing functions.
We also now use Data.Text to write to the file instead of LByteString
which can cause files to remain open.

See https://stackoverflow.com/a/5053188/1360368

@noteed this is a classic example of bad error reporting in
Haskell (and software in general). If the error message here would've
provided us with any contextual information (file name, for example);
debugging this would've been much more straightfoward.

Unfortunately; I've not been able to fix this or to locate the cause
of the error.

The error seems to indeed emanate from Loggers; and it seems to only
occur at runtime, which causes this problem to not surface during
boot or at shutdown.