Fix "resource busy (file is locked)" issue.
This reverts the custom LogCallback to LogFile, but this time using a
greater than 0 log_backup_number. Otherwise, as soon as the
log_file_size limit is reached, the error message appears (when
fast-logger tries to rotate the log with no backup file, presumably).

In addition of increasing the log_backup_number, the log_file_size is
increased to 1MB. In production, I think this could be increased even
further (and we should probably use a real log rotation that adds a
timestamp to files and keep them forever). The original value was much
smaller, but in some way it was a good thing as this made the above
problem easily reproducible.

Some comments about debugging the issue are in
https://github.com/hypered/curiosity/pull/70#issuecomment-1239514419