Add a Core module.
The idea is that

- Data should be pure code, with no STM or IO.
- Core should be for adding an STM layer around Data.
- Runtime should add IO to Core, mainly for logging and calling
  STM.atomically.

Here we can see that `reset` has now proper logs, and `createUser` too.

`handleCommand` is now only responsible (for `CreateUser` for now) to
translate the output into something suitable for `cty` and `cty run` and
`cty repl`.

Similarly, `handleSignup` is reponsible for some additional logging and
translating the output into web pages.