[nix] Filtering sources using a whitelist approach
The Haskell build is quite expensive, we aggressively filter its src
input to make sure we that editor's artifacts and other mutable state
files don't end up triggering a rebuild.

We take advantage of this filtering rework to move back the haskell
sources filtering to the edge: we clean the derivation src before
sending it to cabal2nix projects list. Doing this allow us to apply a
package-specific filter instead of a generic one.