[nix] Integration test: introduce testbench env
This is a first step towards writing an integration test suite.

We set up an environment running the curiosity server behind a Nginx
proxy. This environment is mostly provisioned using the various nix
build artifacts.

While this setup do not provides an isolation level comparable to a
VM-based one, it's much more lightweight.

We had to cut some corners on the Nginx side. We did not find a
satisfactory solution to extract the configuration file from a NixOS
machine description. We're instead using a template text file.

We expose this script via a new target in the top-level default.nix,
via a script file living in `/scripts/integration-tests/` and via a
`nix-shell` bash alias.