Add a registration date to user profile.
This also adds a Twitter username link to the user profile.

Adding a registration date was tricky: it needs to handle bot a
wall-clock time (e.g. to use the demo web site in a realistic way) and a
simulated time (e.g. to craft carefully test scenarios that can be
re-played deterministically). These options are called here Stepping
modes.

We have added three modes, although only two are implemented: Normal,
Stepped, and Mixed.

Those modes are recorded in the state file when that state file is
created (with `cty init`).