Server systemd module: small refactor.
We split the service script in two parts:
  - The service data provisioning is done as part of the PreStart
    script.
  - The server is itself started part of the ExecStart script.

We set the service type to notify to make sure systemd won't consider
the service as ready *before* the Haskell service actually start
listening to its socket. It's the best way to solve some startup
timing issues we faced in the NixOS VM test.

We had to modify a bit the server startup sequence to make it send a
UDP packet to systemd once warp started listening to its socket. Note:
this code will be conceptually no-op when the server isn't executed
via systemd.

More about the systemd notify mechanism here:
- https://www.freedesktop.org/software/systemd/man/sd_notify.html#Description
- https://www.freedesktop.org/software/systemd/man/systemd.service.html#Type=