sources: replace git sources with tarball releases
We were retrieving the smart-design-hs and commence sources through a
git fetcher. Using a git fetcher in a GitHub action would force us to
setup a deployment key to clone those.

These repositories are public and hosted on GitHub, let's retrieve
them as tarballs.

On top of simplifying the CI setup, it'll make the external sources
retrieval process more efficient: tarballs only pack the latest
commit, it doesn't require us to download the full git history of the
dependency.